+++ src/Util/Lifecycle/Method/BoletoBancarioMethodData.php
@@ -49,7 +49,7 @@ class BoletoBancarioMethodData extends AbstractMethodData
 
     public function getInitialState(): bool
     {
-        return false;
+        return true;
     }
 
     public function getMediaFileName(): ?string
@@ -59,11 +59,6 @@ class BoletoBancarioMethodData extends AbstractMethodData
 
     public function validateCapability(MerchantIntegrations $merchantIntegrations): string
     {
-        $product = $merchantIntegrations->getSpecificProduct('PPCP_STANDARD');
-        if ($product !== null && (\in_array($product->getVettingStatus(), [Product::VETTING_STATUS_APPROVED, Product::VETTING_STATUS_SUBSCRIBED], true))) {
-            return self::CAPABILITY_ACTIVE;
-        }
-
-        return self::CAPABILITY_INELIGIBLE;
+        return self::CAPABILITY_ACTIVE;
     }
 }