+++ src/Util/Lifecycle/Method/GiropayMethodData.php
@@ -46,7 +46,7 @@ class GiropayMethodData extends AbstractMethodData
 
     public function getInitialState(): bool
     {
-        return false;
+        return true;
     }
 
     public function getMediaFileName(): ?string
@@ -56,11 +56,6 @@ class GiropayMethodData extends AbstractMethodData
 
     public function validateCapability(MerchantIntegrations $merchantIntegrations): string
     {
-        $product = $merchantIntegrations->getSpecificProduct('PPCP_STANDARD');
-        if ($product !== null && (\in_array($product->getVettingStatus(), [Product::VETTING_STATUS_APPROVED, Product::VETTING_STATUS_SUBSCRIBED], true))) {
-            return self::CAPABILITY_ACTIVE;
-        }
-
-        return self::CAPABILITY_INELIGIBLE;
+        return self::CAPABILITY_ACTIVE;
     }
 }