+++ tests/OrdersApi/Builder/OrderFromCartBuilderTest.php
@@ -28,6 +28,7 @@ use Swag\PayPal\Setting\Exception\PayPalSettingsInvalidException;
 use Swag\PayPal\Setting\Settings;
 use Swag\PayPal\Test\Helper\CartTrait;
 use Swag\PayPal\Test\Helper\ServicesTrait;
+use Swag\PayPal\Test\Mock\CustomIdProviderMock;
 use Swag\PayPal\Test\Mock\EventDispatcherMock;
 use Swag\PayPal\Test\Mock\LoggerMock;
 use Swag\PayPal\Util\PriceFormatter;
@@ -212,11 +213,12 @@ class OrderFromCartBuilderTest extends TestCase
         $priceFormatter = new PriceFormatter();
         $amountProvider = new AmountProvider($priceFormatter);
         $addressProvider = new AddressProvider();
+        $customIdProvider = new CustomIdProviderMock();
 
         return new OrderFromCartBuilder(
             $priceFormatter,
             $systemConfig,
-            new PurchaseUnitProvider($amountProvider, $addressProvider, $systemConfig),
+            new PurchaseUnitProvider($amountProvider, $addressProvider, $customIdProvider, $systemConfig),
             $addressProvider,
             new EventDispatcherMock(),
             new LoggerMock()